chore(Candytar): Moved to own module
