+++ src/components/candytar/Candytar.test.js
